SCAL-646 moved SiM handling loop to separate file, reduced redundancy
