+++ qsub.go
@@ -4,16 +4,16 @@ import (
 	"fmt"
 	"regexp"
 	"strings"
-
-	"github.com/scalarm/scalarm_workers_manager/logger"
 )
 
-type QsubFacade struct{}
+type QsubFacade struct {
+	PLGridFacade
+}
 
 //receives command to execute
 //executes command, extracts resource ID
 //returns new job ID
-func (qf QsubFacade) prepareResource(ids, command string) (string, error) {
+func (qf QsubFacade) PrepareResource(ids, command string) (string, error) {
 	stringOutput, err := execute(ids, command)
 	if err != nil {
 		return "", fmt.Errorf(stringOutput)
@@ -43,7 +43,7 @@ func (qf QsubFacade) StatusCheck() ([]string, error) {
 //receives job ID
 //checks resource state based on job state
 //returns resource state
-func (qf QsubFacade) resourceStatus(statusArray []string, jobID string) (string, error) {
+func (qf QsubFacade) ResourceStatus(statusArray []string, jobID string) (string, error) {
 	if jobID == "" {
 		return "available", nil
 	}
@@ -103,93 +103,4 @@ func (qf QsubFacade) resourceStatus(statusArray []string, jobID string) (string,
 	}
 	// no such jobID
 	return "released", nil
-}
-
-//receives sm_record, ExperimentManager connector and infrastructure name
-//decides about action on sm and its resources
-//returns nothing
-func (qf QsubFacade) HandleSM(sm_record *Sm_record, emc *ExperimentManagerConnector, infrastructure string, statusArray []string) {
-	resource_status, err := qf.resourceStatus(statusArray, sm_record.Job_id)
-	if err != nil {
-		sm_record.Error_log = err.Error()
-		sm_record.Resource_status = "error"
-		return
-	}
-
-	ids := sm_record.GetIDs()
-
-	logger.Debug("%v Sm_record state: %v ", ids, sm_record.State)
-	logger.Debug("%v Resource status: %v ", ids, resource_status)
-
-	if sm_record.Cmd_to_execute_code != "" {
-		logger.Info("%v Command to execute: %v ", ids, sm_record.Cmd_to_execute_code)
-	}
-
-	defer func() {
-		sm_record.Cmd_to_execute = ""
-		sm_record.Cmd_to_execute_code = ""
-	}()
-
-	if (sm_record.Cmd_to_execute_code == "prepare_resource" && resource_status == "available") || sm_record.Cmd_to_execute_code == "restart" {
-
-		if _, err := RepetitiveCaller(
-			func() (interface{}, error) {
-				return nil, emc.GetSimulationManagerCode(sm_record.Id, infrastructure)
-			},
-			nil,
-			"GetSimulationManagerCode",
-		); err != nil {
-			logger.Fatal("Unable to get simulation manager code")
-		}
-
-		//extract first zip
-		err := extract("sources_"+sm_record.Id+".zip", ".")
-		if err != nil {
-			sm_record.Error_log = err.Error()
-			sm_record.Resource_status = "error"
-			return
-		}
-		//move second zip one directory up
-		_, err = executeSilent("mv scalarm_simulation_manager_code_" + sm_record.Sm_uuid + "/* .")
-		if err != nil {
-			sm_record.Error_log = err.Error()
-			sm_record.Resource_status = "error"
-			return
-		}
-		//remove both zips and catalog left from first unzip
-		_, err = executeSilent("rm -rf  sources_" + sm_record.Id + ".zip scalarm_simulation_manager_code_" + sm_record.Sm_uuid)
-		if err != nil {
-			sm_record.Error_log = err.Error()
-			sm_record.Resource_status = "error"
-			return
-		}
-		logger.Debug("%v Code files extracted", ids)
-
-		//run command
-		jobID, err := qf.prepareResource(ids, sm_record.Cmd_to_execute)
-		if err != nil {
-			sm_record.Error_log = err.Error()
-			sm_record.Resource_status = "error"
-			return
-		}
-		logger.Info("%v Assigned job_id: %v", ids, jobID)
-		sm_record.Job_id = jobID
-
-	} else if sm_record.Cmd_to_execute_code == "stop" {
-
-		stringOutput, err := execute(ids, sm_record.Cmd_to_execute)
-		if err != nil {
-			sm_record.Error_log = stringOutput
-			sm_record.Resource_status = "error"
-			return
-		}
-
-	} else if sm_record.Cmd_to_execute_code == "get_log" {
-
-		stringOutput, _ := execute(ids, sm_record.Cmd_to_execute)
-		sm_record.Error_log = stringOutput
-
-	}
-
-	sm_record.Resource_status = resource_status
 }
